added additional code play-pause file
called a custom html5 tag called "video-src" which goes through all the seasons currently in the front-end html tags view for meditation under the season class , and selects the video file for each season e.g. winter, spring etc , and upon play where the user selects a season it fills the background with that video file.